requirements.txt
ollama
langchain-community
unstructured
unstructured[md]
sentence-transformers
chromadb